% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.icd.R
\name{group.icd}
\alias{group.icd}
\alias{chapter.icd}
\alias{subchapter.icd}
\title{Higher level groupings for ICD.10 codes}
\usage{
chapter.icd(x)

subchapter.icd(x)
}
\arguments{
\item{x}{\code{character} An ICD code, for example 'R85.1'.}
}
\description{
For a  variable containing ICD-10 Codes what are the associated higher level groupings? \cr

The International Classification of Diseases, Tenth Revision (ICD-10) is a comprehensive system of codes used to classify and standardize medical diagnoses, symptoms, and procedures. Developed by the World Health Organization (WHO), ICD-10 provides a universal language for healthcare professionals to communicate and record medical information.
Due to underlying assumptions of research questions and nuances in healthcare, a method for grouping codes in one scenario may not be suitable in another scenario.  As such, grouping ICD-10 codes into higher levels is frequently inconsistent. This package offers several grouping levels that may be a useful starting point for researchers.
}
\note{
Can pass a vector. \cr

See also
https://icd.who.int/browse10/2019/en
https://www.icd10data.com/ICD10CM/Codes

@example

df <- data.frame(ID = c( 1,2,3,4,5), my.var = c('A3.0', 'R85.1', 'F64', 'U10.2', 'O9A'))

df$chapter <- chapter.icd(df$my.var)
}
