% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt.rci.R
\name{jt.rci}
\alias{jt.rci}
\title{Reliable Change Index}
\source{
Refer to the article by Jacobson & Truax (1991) for a complete explanation:

\href{https://pubmed.ncbi.nlm.nih.gov/2002127/}{Clinical significance: a statistical approach to defining meaningful change in psychotherapy research} \cr

\cr
}
\usage{
jt.rci(
  rxx,
  wide.df,
  pre.vec,
  post.vec,
  sd_ref,
  NA_handling = "ignore NAs",
  Categories = "cat.3",
  unused_factor_drop = FALSE,
  min_to_use,
  max_to_use,
  higherIsBetter.num = 0,
  recovery_thrshld = FALSE,
  thrshld_ref,
  thrshld_quantile = 5,
  conf.lvl = 0.95,
  sd_ref.cite,
  rxx.cite,
  thrshld_ref.cite,
  title,
  subtitle,
  xlabel,
  ylabel,
  caption,
  colour_palette,
  colour.vec
)
}
\arguments{
\item{rxx}{The instrument test-retest reliability (typically ICC or Cronbach's alpha etc.)}

\item{wide.df}{A wide data frame, with pre and post instrument scores for each paired/matched observation}

\item{pre.vec}{Vector defining sample pre instrument scores}

\item{post.vec}{Vector defining sample post instrument scores}

\item{sd_ref}{A reference/population standard deviation, if available, else the standard deviation of the pre-test scores is used.}

\item{NA_handling}{How should NAs be handled? Default is: \code{'ignore NAs'} - and calculates the SD of all pre-test scores available.  If set to \code{'use complete obs only'}, function will calculate the SD of observations from the set of pre-test scores for which there are also post-test scores. Note that patient data for only pre or post time points can be used to calculate RCI or be plotted on the chart, therefore this argument is only for specifying how to calculate the SD from the pre-test scores.}

\item{Categories}{How should participants be grouped? Options are: \code{'cat.3'}, \code{'cat.5'}, \code{'JT.1'}, \code{'JT.2'}, \code{'JT.4'}.}

\item{unused_factor_drop}{Should unrepresented categorizations be dropped in plot and tabled output?}

\item{min_to_use}{Minimum possible value of the instrument? If not provided the least positive/most negative value from both pre and post scores will be used.}

\item{max_to_use}{Maximum possible value of the instrument? If not provided the most positive value from both pre and post scores will be used.}

\item{higherIsBetter.num}{Are higher scores healthier? '1' - higher scores indicative of better health, '0' - higher scores indicative of worse health. (Defaults to '0' ).}

\item{recovery_thrshld}{Should the recovery threshold information be included? If Categorisation requires the threshold, the recovery argument is ignored, and the threshold is plotted regardless.}

\item{thrshld_ref}{Insert a value for a threshold that defines \emph{'case-ness'}. I.e. A threshold that is indicative of diagnosis, and not surpassing this threshold, a person would be deemed asymptomatic or recovered. If not supplied the function will impute the value as the healthiest quintile of the pre-test scores. Note that for \code{higherIsBetter.num = 0} (default) the lowest quintile is used. NA_handling also applies here and will tell the function whether to use all the pre-test scores (\code{'ignore NAs'}) or only those for which there are also post-test scores (\code{'use complete obs only'})}

\item{thrshld_quantile}{An argument to specify which quantile to use to define recovery (defaults to \code{'5'}, i.e. quintiles).}

\item{conf.lvl}{With what confidence level should false positives and negatives be screened? Default is \code{0.95}.}

\item{sd_ref.cite}{\code{sd_ref} source?}

\item{rxx.cite}{\code{rxx} source?}

\item{thrshld_ref.cite}{\code{thrshld_ref} source?}

\item{title}{ggplot \code{labs()} argument \code{title}}

\item{subtitle}{ggplot \code{labs()} argument \code{subtitle}}

\item{xlabel}{ggplot \code{labs()} argument \code{xlabel}}

\item{ylabel}{ggplot \code{labs()} argument \code{ylabel}}

\item{caption}{ggplot \code{labs()} argument \code{caption}}

\item{colour_palette}{Provide a vector at least three (or 5 for larger classification) colours for chart. Otherwise defaults are used. This Will directly place the colour palette into the \code{ggplot()}. It is quick if only making a few charts. Use argument \code{colour.vec} if making many charts.}

\item{colour.vec}{Supply a vector of 5 colors. If using the three-way categorization (default), the second, fourth and fifth positions in the vector will be used. These reflect similar factor levels in differing categorisations. Recommended over \code{colour_palette}.}
}
\value{
Returns a list including:
\itemize{
\item the RCI
\item a ggplot using ~ \code{geom_point()}
\item a data frame of proportions
\item the the reference sd/sd used
}

\cr
}
\description{
Calculate the Reliable Change Index (RCI), and produce a corresponding plot in keeping with \href{https://pubmed.ncbi.nlm.nih.gov/2002127/}{Jacobson & Truax's 1991 article}. \cr

While a p value tells us about the probability of observing a change of certain size, and a Standardised mean difference
(SMD ~ Cohen's d/Hedge's g) indicates
a probabilistic or statistical magnitude, a Reliable Change Index (RCI) discerns whether a change score measurement is a true positive or negative, according to; a pre-specified level of confidence,
the error of an instrument, and a normative distribution.

\cr

Extending the RCI to clinical significance suggests that changes in instrument scores larger than an RCI, would be perceptible
by an individual/patient, therapist, or significant other.
It is a common metric to estimate a threshold required to meet levels of clinical significance in psychometric studies, and
compatible with the concept of a Minimally Clinically Important Difference (MCID).

\cr
}
\details{
The function \code{jt.rci()} manages psychometric scores at two time-points in line with Jacobson & Truax's foundational paper outlining the Reliable Change Index (RCI).
At a minimum, the function requires pre & post psychometric instrument scores from a paired or matched sample, and Jacobson and Truaux's value: \code{rxx} (usually ~ ICC or Cronbach's alpha).
If no argument is given for \code{sd_ref}, it is calculated from the pre-test scores.
The function will output a plot, an RCI, and a table detailing the proportions and numbers of individuals who may have improved, stayed the same, or deteriorated.

The RCI is the real value of an instrument scale, with which we can say with a certain level of confidence (default is 95\%),
that a clinically significant level of change has occurred. The RCI separates a sample of participants into those who have had truly positive improvement, deterioration, or those who who have had indeterminate change.
Therefore it is important to also report the percentages who reliably improved, stayed the same, and reliably deteriorated.
Participants who fall within 1.96 (\code{conf.lvl = 0.95}) RC of no change (these observations are located the diagonal bounds of the plots produced by the \code{jt.rci()} function) are considered false positives/false negatives.

\cr
}
\note{
\itemize{
\item Test/retest reliabilities are often included in validation studies. Try to use a reliability that was observed from a sample characteristically similar to the one you are investigating (Example: adults vs children )
\item For analyzing the RCI over multiple time points see \href{https://dx.doi.org/10.1002/mpr.1906}{Morgan-Lopez et. al. (2022)}
}

\cr
}
\examples{


# Examples ----

library(tidyverse)

# using the sample data in `rci.data` dataframe run the following code.
# Execution will produce some values, an RCI, a plot, a table, and the sd used (also printed in default caption)
# The function should print a list in the console, including an RCI (first object), a plot (second object - in the 'Plots' tab), the table (third object) of instances with proportions of individuals in each group, and the sd used (fourth object)

jt.rci(

 0.88,
 rci.data,
 pre,
 post

)



## Calling individual outputs: ----

#  in this code the output ( a list) is placed into an object (my_jt.rci_object) and saved in the  work environment,

my_jt.rci_object <- jt.rci(

  0.88,

  rci.data,
  pre,
  post

)

# access the individual outputs using the following:

my_jt.rci_object$rci   # the estimated RCI for this data
my_jt.rci_object$plot   # plot
my_jt.rci_object$table   # tabled results
my_jt.rci_object$sd_used   # sd used

# or alternatively...

my_jt.rci_object[[1]]   # the estimated RCI for this data
my_jt.rci_object[[2]]   # plot
my_jt.rci_object[[3]]   # tabled results
# ...



# the RCI is numeric i.e: it can be treated as a number

my_jt.rci_object[[1]]*15                   # multiply
round( my_jt.rci_object[[1]] , 2)          # rounding decimals

# the plot is a ggplot object i.e: it can be treated as a ggplot2 object and exported as such

my_jt.rci_object[[2]] + ggplot2::theme_minimal()
my_jt.rci_object[[2]] + ggplot2::theme(legend.position = "top")

# the tabled results can be inspected as a dataframe:

table_df <- my_jt.rci_object[[3]]         # save the table (as dataframe)
View(table_df)                            # view the dataframe

# change the order of the columns

rearranged_table_df <- table_df \%>\% select(c(1,3,2), everything())
View(rearranged_table_df)



# Other options: ----

##  NA_handling ----

# in the first example, the SD used to calculate the output was printed in the caption of the plot and printed in the bottom of the screen
# it can also be retrieved:

my_jt.rci_object$sd_used

# Users can tell the function to only use pre.vec-test scores for which there are corresponding post-test scores by setting the argument 'NA_handling' = 'use complete obs only'
# The sd_used should have changed slightly in these two examples, as the pre and post scores in rci.data contain missing values

jt.rci(

  0.88,
  rci.data,
  pre,
  post,
  NA_handling = 'use complete obs only',

)



## Recovery threshold: ----

# display the recovery threshold using  recovery_thrshld = TRUE,

jt.rci(

  0.78,
  rci.data,
  pre,
  post,
  recovery_thrshld = TRUE

)

# Alternatively, using Categories = 5 (or other arguments that depend on the recovery threshold) will ignore recovery_thrshld argument and include the threshold as mandatory

jt.rci(

  0.78,
  rci.data,
  pre,
  post,
  Categories = 'cat.5',
  recovery_thrshld = FALSE     # ignores this argument

)



## Supply a threshold of recovery: ----

# supply a threshold from the literature

jt.rci(

  0.78,
  rci.data,
  pre,
  post,
  Categories = 'cat.3',
  thrshld_ref = 8,
  recovery_thrshld = TRUE,
  higherIsBetter.num = 1

)



## higherIsBetter.num: ----

# If higher scores are indicative of better health higherIsBetter.num should be switched to '1'

jt.rci(

  0.98,
  rci.data,
  pre,
  post,
  Categories = 'cat.5',
  higherIsBetter.num = 1

)


## Unrepresented levels: ----

# If a poor/low rxx is selected, it is unlikely that individuals will fall into the 'Unreliably Recovered' category  (5 levels). If so there will be no colour placed next to this level in the chart, as it is not plotted in the plot area.
# If it is preferred that these levels are not reported (though not recommended), user can set 'unused_factor_drop' to 'TRUE'

jt.rci(

  0.78,
  rci.data,
  pre,
  post,
  Categories = 'cat.5',
  higherIsBetter.num = 1 ,
  unused_factor_drop = TRUE

)



## Supplied sd: ----

jt.rci(

  0.88,
  rci.data,
  pre,
  post,
  sd_ref = 8.902,
  sd_ref.cite = '(My Reliable Sauce)',

)



## Citations: ----

jt.rci(

  0.88,

  rci.data,
  pre,
  post,
  sd_ref.cite = ' - My Reliable Sauce',
  rxx.cite = ' (Scooby & Doo (2025))',
  thrshld_ref.cite = '(~ a Validation Study)',
  recovery_thrshld = TRUE
)



## Plot area: ----

# It's usually best to place the full range of the psychometric instrument on the plot
# Specify these using the min_to_use and max_to_use arguments
# For an example suppose the true maximums and minimums of a psychometric scale are defined by minima and maxima:

minima = plyr::round_any( min(c( rci.data[ , 2], rci.data[ , 3] ), na.rm = TRUE) - 5 , 1,  floor) # add padding (10) to the lower limit
maxima = plyr::round_any( max(c( rci.data[ , 2], rci.data[ , 3] ), na.rm = TRUE) + 5 , 1,  ceiling) # add padding (5) to the upper limit


jt.rci(

  0.88,
  rci.data,
  pre,
  post,
  min_to_use =  minima,
  max_to_use = maxima,

)



## Colours: ----

# Using the colour_palette argument you can intuitively place the colours in a vector to match the displayed output
# Provide exactly the colours you prefer in a vector

jt.rci(

  0.78,
  rci.data,
  pre,
  post,
  Categories = 'cat.3',
  colour_palette = c('blue', 'darkgreen', 'red'),

)

# Note that if not enough colours are provided errors ensue:
# "! Insufficient values in manual scale. 5 needed but only 3 provided"

jt.rci(

  0.78,
  rci.data,
  pre,
  post,
  Categories = 'cat.5',
  colour_palette = c('blue', 'darkgreen', 'red'),
)


# the colour.vec is a better way to choose preferred colours and will reflect categorisations consistently over many trials
# Note that'colour_palette' is a direct value placed in a ggplot and will override colour.vec if both are specified in the function.

jt.rci(

  0.70,
  rci.data,
  pre,
  post,
  Categories = 'cat.5',
  colour.vec = c('blue', 'darkgreen', 'red', 'yellow', 'brown'),
  unused_factor_drop = TRUE

)



## Plot labels: ----

jt.rci(

  0.88,
  rci.data,
  pre,
  post,

  title = 'My title',
  subtitle = 'Pre-post scores of Kessler-10 among adults',
  caption = '',
  xlabel = 'Pre K-10',
  ylabel = 'Post K-10'

)



# JT Examples ----

# DAS example

thrshld_ref.DAS = round(jt_a(jt.data$DAS.Pretest, higherIsBetter = 1),1)

jt.rci(rxx = 0.96,
       jt.data,
       DAS.Pretest,
       DAS.Posttest,
       sd_ref = 17.8,
       min_to_use = 40,
       max_to_use = 140,
       higherIsBetter.num = 1,
       thrshld_ref = thrshld_ref.DAS,
       thrshld_ref.cite = paste0(" (Jacobson & Truax\'s value: 'a')" ),
       sd_ref.cite = ' (Spanier, 1976)' ,
       rxx.cite = ' (Spanier, 1976)',
       Categories = 'jt.4'
)


# GDS example

thrshld_ref.GDS = round(jt_c(jt.data$GDS.Pretest, m0 = 50), 1)

jt.rci(rxx = 0.75,
       jt.data,
       GDS.Pretest,
       GDS.Posttest,
       higherIsBetter.num = 0,
       thrshld_ref = thrshld_ref.GDS,
       thrshld_ref.cite = paste0(" (Jacobson & Truax\'s value: 'c')" ),
       rxx.cite = ' (average correlation of MSI subscales - Snyder, 1976)',
       Categories = 'jt.4'
)




}
\seealso{
\itemize{
\item\href{https://dx.doi.org/10.1017/s1754470x22000484}{Reliable change and the reliable change index: still useful after all these years?}
\item\href{https://doi.org/10.1207/S15327752JPA8001_18}{Starting at the beginning: an introduction to coefficient alpha and internal consistency}
}

\cr
}
\keyword{'MCID'}
\keyword{'RCC'}
\keyword{'RCI'}
\keyword{'psychometrics'}
