% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_cutoffs.R
\name{jt_cutoffs}
\alias{jt_cutoffs}
\alias{jt_a}
\alias{jt_b}
\alias{jt_c}
\title{Cutoffs functions: jt.a, jt.b, jt.c}
\usage{
jt_a(vec, m1, s1, higherIsBetter)

jt_b(vec, m0, s1, higherIsBetter)

jt_c(m0, s0, m1, s1, equal.variance)
}
\arguments{
\item{vec}{Instrument scores for sample of interest.}

\item{m1}{Mean of pretest experimental or pretest control groups. If \code{vec} is supplied, \code{m1} will be calculated automatically.}

\item{s1}{Mean of pretest experimental or pretest control groups. If \code{vec} is supplied, \code{s1} will be calculated automatically.}

\item{m0}{Mean instrument score among a well functioning or normal population.}

\item{s0}{Standard deviation of instrument score among a well functioning or normal population.}

\item{equal.variance}{If \code{TRUE}, only \code{m0} & \code{m1} (or \code{vec}) are required. If \code{FALSE}, function requires: \code{m0}, \code{s0}, \code{m1} & \code{s1}. Defaults to \code{TRUE}.}
}
\description{
Cutoffs described by \href{https://pubmed.ncbi.nlm.nih.gov/2002127/}{Jacobson and Truax (1991)}
}
\note{
The functions \code{jt.a()}, \code{jt.b()}, and \code{jt.c()} each return a threshold which specify recovery in a population. \cr

\itemize{
\item \code{jt.a()} has the advantage of not requiring referent population norms. When distributions of well and unwell populations are overlapping, \code{jt_a()} can potentially be quite conservative. However, when distributions are non-overlapping, \code{jt_a()} will not be conservative enough.
\item \code{jt.b()} & \code{jt.c()} both require artefacts from normal populations
\item  if distributions are overlapping then threshold \code{jt.b()} is rather lenient. If non-overlapping, \code{jt.b()} can potentially be too stringent, but provides confidence of recovery.
} \cr

See \href{(https://pubmed.ncbi.nlm.nih.gov/2002127/)}{Jacobson & Truax's 1991} paper outlining the Reliable Change Index (RCI).

#' @example rci_helpers_examples.R
}
