% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCI_helpers.R
\name{rci_helpers}
\alias{rci_helpers}
\alias{SEm}
\alias{sdiff}
\alias{RC}
\alias{RC2}
\alias{RCI}
\title{Helper functions for investigating the Reliable Change Index (RCI)}
\source{
See \href{(https://pubmed.ncbi.nlm.nih.gov/2002127/)}{Jacobson & Truax (1991)} outlining the Reliable Change Index (RCI). \cr
}
\usage{
SEm(sd_norm, rxx)

sdiff(SEm)

RC(pre_mean, post_mean, sdiff)

RC2(pre_mean, post_mean, sd_norm, rxx)

RCI(sdiff, conf.lvl)
}
\arguments{
\item{sd_norm}{A reference/population standard deviation}

\item{rxx}{The instrument test-retest reliability (typically ICC or Cronbach's alpha etc.)}

\item{SEm}{standard error of measurement}

\item{pre_mean}{A mean of pre scores}

\item{post_mean}{A mean of post scores}

\item{sdiff}{standard error of measurement of differences}
}
\description{
Smaller functions that can be used quickly to calculate quantities related to the Reliable Change Index (RCI) \cr
}
\note{
The following functions describe various steps to calculating the RCI:

\itemize{
\item SEm: Standard error of measurement (In Jacobson & Truax (1991) this is know as SE). For SEm - sd_norm is the Standard deviation of a normal population, or control group, or pre-treatment experimental group, rxx is the test- retest reliability, or internal consistency (E.g. ICC, Cronbach's alpha)
\item Sdiff: standard error of the difference between pre and post scores. Describes the spread of the distribution (sd) of change scores, that would be expected if no actual change had occurred. An RC larger than 1 .96 would be unlikely to occur (p < .05) without actual change.
\item RC: is reminiscent of an SMD but slightly different. The denominators for both an SMD and RC are spreads of data. For d it is of the change distribution, whereas for RC it is of the distribution for which no change has occurred. If RC is greater than 1.96, it is a rare observation at the 95\% confidence level.
\item RC2: is the same as RC, but takes the rxx, and sd_norm directly, rather than the computed sdiff
\item RCI: the level required for clinical significance on the same scale as the instrument in question.
} \cr
}
\examples{



SE <- SEm(7.5, 0.8)

sdiff <- sdiff(SE)

RC(47.5, 32.5, sdiff)

RC2(47.5, 32.5, 7.5, 0.8)

RCI(17.8, 0.96)
}
